+++ src/main/java/deco2800/skyfall/managers/database/queries.sql
@@ -1,6 +1,6 @@
 CREATE TABLE SAVES
 (
-    save_id       int not null,
+    save_id       bigint not null,
     data clob,
     PRIMARY KEY (save_id)
 );
@@ -8,8 +8,8 @@ CREATE TABLE SAVES
 
 CREATE TABLE WORLDS
 (
-    save_id int not null ,
-    world_id int not null ,
+    save_id bigint not null ,
+    world_id bigint not null ,
     is_current_world boolean,
     data clob,
     primary key (world_id) ,
@@ -19,8 +19,8 @@ CREATE TABLE WORLDS
 
 CREATE TABLE MAIN_CHARACTER
 (
-    character_id        int NOT NULL,
-    save_id             int not null ,
+    character_id        bigint NOT NULL,
+    save_id             bigint not null ,
     data blob,
     PRIMARY KEY (character_id),
     FOREIGN KEY (save_id) references SAVES (save_id)
@@ -28,11 +28,11 @@ CREATE TABLE MAIN_CHARACTER
 
 CREATE TABLE NODES
 (
-    world_id int not null ,
-    x_pos int not null ,
-    y_pos int not null ,
+    world_id bigint not null ,
+    x_pos bigint not null ,
+    y_pos bigint not null ,
     data CLOB,
-    node_id int,
+    node_id bigint,
     primary key (node_id),
     foreign key (world_id) references WORLDS(world_id)
 );
@@ -40,8 +40,8 @@ CREATE TABLE NODES
 
 CREATE TABLE EDGES
 (
-    node_one_id int,
-    node_two_id int,
+    node_one_id bigint,
+    node_two_id bigint,
     primary key (node_one_id, node_two_id),
     foreign key (node_two_id) references NODES(node_id),
     foreign key (node_one_id) references NODES(node_id)
@@ -50,8 +50,8 @@ CREATE TABLE EDGES
 
 CREATE table BIOMES
 (
-    biome_id       int not null ,
-    world_id       int not null ,
+    biome_id       bigint not null ,
+    world_id       bigint not null ,
     biome_type     CLOB ,
     tile_generator blob not null ,
     data CLOB,
@@ -61,7 +61,7 @@ CREATE table BIOMES
 
 CREATE TABLE CHUNKS
 (
-    world_id int not null ,
+    world_id bigint not null ,
     x int not null ,
     y int not null ,
     data CLOB,
@@ -76,8 +76,12 @@ CREATE TABLE ENTITIES
     y int not null ,
     chunk_x int not null,
     chunk_y int not null,
-    world_id int not null,
+    world_id bigint not null,
     data clob,
     PRIMARY KEY (world_id,chunk_x, chunk_y, x, y),
     foreign key (world_id, chunk_x, chunk_y) references CHUNKS(world_id, x, y)
 )
+
+
+
+
